+++ src/main/java/com/twilio/rest/verify/v2/VerificationAttemptsSummaryFetcher.java
@@ -21,7 +21,7 @@ import com.twilio.rest.Domains;
 import java.time.ZonedDateTime;
 
 public class VerificationAttemptsSummaryFetcher extends Fetcher<VerificationAttemptsSummary> {
-    private String serviceSid;
+    private String verifyServiceSid;
     private ZonedDateTime dateCreatedAfter;
     private ZonedDateTime dateCreatedBefore;
     private String country;
@@ -32,12 +32,12 @@ public class VerificationAttemptsSummaryFetcher extends Fetcher<VerificationAtte
      * Filter used to consider only Verification Attempts of the given verify
      * service on the summary aggregation..
      *
-     * @param serviceSid Filter the verification attempts considered on the summary
-     *                   by verify service.
+     * @param verifyServiceSid Filter the verification attempts considered on the
+     *                         summary by verify service.
      * @return this
      */
-    public VerificationAttemptsSummaryFetcher setServiceSid(final String serviceSid) {
-        this.serviceSid = serviceSid;
+    public VerificationAttemptsSummaryFetcher setVerifyServiceSid(final String verifyServiceSid) {
+        this.verifyServiceSid = verifyServiceSid;
         return this;
     }
 
@@ -143,8 +143,8 @@ public class VerificationAttemptsSummaryFetcher extends Fetcher<VerificationAtte
      * @param request Request to add query string arguments to
      */
     private void addQueryParams(final Request request) {
-        if (serviceSid != null) {
-            request.addQueryParam("ServiceSid", serviceSid);
+        if (verifyServiceSid != null) {
+            request.addQueryParam("VerifyServiceSid", verifyServiceSid);
         }
 
         if (dateCreatedAfter != null) {