+++ src/main/java/com/twilio/rest/verify/v2/service/VerificationCheckCreator.java
@@ -19,23 +19,20 @@ import com.twilio.rest.Domains;
 
 public class VerificationCheckCreator extends Creator<VerificationCheck> {
     private final String pathServiceSid;
-    private final String code;
     private String to;
     private String verificationSid;
     private String amount;
     private String payee;
+    private String code;
 
     /**
      * Construct a new VerificationCheckCreator.
      *
      * @param pathServiceSid The SID of the verification Service to create the
      *                       resource under
-     * @param code The verification string
      */
-    public VerificationCheckCreator(final String pathServiceSid,
-                                    final String code) {
+    public VerificationCheckCreator(final String pathServiceSid) {
         this.pathServiceSid = pathServiceSid;
-        this.code = code;
     }
 
     /**
@@ -91,6 +88,17 @@ public class VerificationCheckCreator extends Creator<VerificationCheck> {
     }
 
     /**
+     * The 4-10 character string being verified..
+     *
+     * @param code The verification string
+     * @return this
+     */
+    public VerificationCheckCreator setCode(final String code) {
+        this.code = code;
+        return this;
+    }
+
+    /**
      * Make the request to the Twilio API to perform the create.
      *
      * @param client TwilioRestClient with which to make the request
@@ -127,10 +135,6 @@ public class VerificationCheckCreator extends Creator<VerificationCheck> {
      * @param request Request to add post params to
      */
     private void addPostParams(final Request request) {
-        if (code != null) {
-            request.addPostParam("Code", code);
-        }
-
         if (to != null) {
             request.addPostParam("To", to);
         }
@@ -145,6 +149,10 @@ public class VerificationCheckCreator extends Creator<VerificationCheck> {
 
         if (payee != null) {
             request.addPostParam("Payee", payee);
+        }
+
+        if (code != null) {
+            request.addPostParam("Code", code);
         }
     }
 }