+++ CHANGES.md
@@ -1,6 +1,25 @@
 twilio-java changelog
 =====================
 
+[2022-07-13] Version 8.33.0
+---------------------------
+**Library - Test**
+- [PR #699](https://github.com/twilio/twilio-java/pull/699): Adding misc as PR type. Thanks to [@rakatyal](https://github.com/rakatyal)!
+
+**Conversations**
+- Allowed to use `identity` as part of Participant's resource **(breaking change)**
+
+**Lookups**
+- Remove `enhanced_line_type` from the lookup response **(breaking change)**
+
+**Supersim**
+- Add support for `sim_ip_addresses` resource to helper libraries
+
+**Verify**
+- Changed summary param `service_sid` to `verify_service_sid` to be consistent with list attempts API **(breaking change)**
+- Make `code` optional on Verification check to support `sna` attempts.
+
+
 [2022-06-29] Version 8.32.0
 ---------------------------
 **Api**